Implementing repository pattern for monthly billings

+++ src/Application/Exceptions/MonthlyBillingAlreadyOpenedException.cs
@@ -4,6 +4,6 @@ namespace Application.Exceptions;
 
 public sealed class MonthlyBillingAlreadyOpenedException : SmugetException
 {
-    public MonthlyBillingAlreadyOpenedException(byte month, ushort year)
+    public MonthlyBillingAlreadyOpenedException(int month, int year)
         : base($"You already have opened monthly billing for `{month}/{year}`.") { }
 }
