+++ src/Domain/Exceptions/InvalidMonthException.cs
@@ -2,9 +2,9 @@ namespace Domain.Exceptions;
 
 public sealed class InvalidMonthException : SmugetException
 {
-    public byte Value { get; }
+    public int Value { get; }
 
-    public InvalidMonthException(byte value)
+    public InvalidMonthException(int value)
         : base($"Month must be an integer number from 1 to 12. Passed: {value}.")
     {
         Value = value;
